feat(enterprise): update IndustryTypeEnum
