+++ domain/enum/IndustryTypeEnum.ts
@@ -1,0 +1,14 @@
+export enum IndustryTypeEnum {
+  Technology = "Technology",
+  Healthcare = "Healthcare",
+  Finance = "Finance",
+  Education = "Education",
+  Retail = "Retail",
+  Manufacturing = "Manufacturing",
+  Energy = "Energy",
+  RealEstate = "Real Estate",
+  Hospitality = "Hospitality",
+  Transportation = "Transportation",
+  Agriculture = "Agriculture",
+  Entertainment = "Entertainment",
+}
