+++ domain/types/IndustryType.ts
@@ -1,19 +1,5 @@
 import { EnterpriseIndustryTypeError } from "../errors/EnterpriseIndustryTypeError.ts";
-
-export enum IndustryTypeEnum {
-  Technology = "Technology",
-  Healthcare = "Healthcare",
-  Finance = "Finance",
-  Education = "Education",
-  Retail = "Retail",
-  Manufacturing = "Manufacturing",
-  Energy = "Energy",
-  RealEstate = "Real Estate",
-  Hospitality = "Hospitality",
-  Transportation = "Transportation",
-  Agriculture = "Agriculture",
-  Entertainment = "Entertainment",
-}
+import { IndustryTypeEnum } from "../enum/IndustryTypeEnum.ts";
 
 export class IndustryType {
   private constructor(public readonly value: string) {}
